feat: support custom user resolution
