+++ src/UserstampsScope.php
@@ -5,7 +5,6 @@ namespace Mattiverse\Userstamps;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Scope;
-use Illuminate\Support\Facades\Auth;
 
 class UserstampsScope implements Scope
 {
@@ -30,22 +29,22 @@ class UserstampsScope implements Scope
     public function extend(Builder $builder)
     {
         $builder->macro('updateWithUserstamps', function (Builder $builder, $values) {
-            if (! $builder->getModel()->isUserstamping() || is_null(Auth::id())) {
+            if (! $builder->getModel()->isUserstamping() || is_null(Userstamps::getUserId())) {
                 return $builder->update($values);
             }
 
-            $values[$builder->getModel()->getUpdatedByColumn()] = Auth::id();
+            $values[$builder->getModel()->getUpdatedByColumn()] = Userstamps::getUserId();
 
             return $builder->update($values);
         });
 
         $builder->macro('deleteWithUserstamps', function (Builder $builder) {
-            if (! $builder->getModel()->isUserstamping() || is_null(Auth::id())) {
+            if (! $builder->getModel()->isUserstamping() || is_null(Userstamps::getUserId())) {
                 return $builder->delete();
             }
 
             $builder->update([
-                $builder->getModel()->getDeletedByColumn() => Auth::id(),
+                $builder->getModel()->getDeletedByColumn() => Userstamps::getUserId(),
             ]);
 
             return $builder->delete();
